+++ src/ui/messages/message-type-list/message-type-list-view.js
@@ -1,4 +1,4 @@
-/**
+/*
  *
  * @class
  * @extends Layer.UI.Component